+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On `ConvertStringUtils` and `ChessGame`, fixed not allowing king promotion even the variant is suicide or giveaway.
 - On `ChessGame.getCapturedPieces`, fixed the captured piece going negative when piece is promoted, but negative number
   ignored and removed.
+- On `ChessGame.makeMoveAll*` methods, changed `string.split(" ")` to `string.isEmpty() ? new String[0] : string.split("\\s+")`.
+- On `ChessGame.internalRemakeMove`, added checking variation index is negative.
 
 ### Performance